PR feedback: fix tutorial folder & README.md
