+++ metaflow/plugins/kfp/README.md
@@ -1,7 +1,5 @@
 ### Commands used and References
 
-###### Last Updated: Aug 12th:
-
 #### Steps to run Metaflow on Kubeflow: 
 
 #### Installing the correct forked version of Metaflow AND KFP
@@ -9,19 +7,20 @@ Currently, to run Metaflow on our internal KFP, you need to install the correct 
 
 To install KFP:
 ```
-    - git clone -b alexla/AIP-1676 https://github.com/alexlatchford/pipelines
-    - cd pipelines/sdk/python
-    - pip3 install -e .
+pip3 install 'git+https://github.com/alexlatchford/pipelines@alexla/AIP-1676#egg=kfp&subdirectory=sdk/python'
 ```
-To install Metaflow: `pip3 install --user --upgrade git+https://github.com/zillow/metaflow.git@feature/kfp`
+
+To install Metaflow: 
+```
+pip3 install --user --upgrade git+https://github.com/zillow/metaflow.git@feature/kfp
+```
 
 ##### Option 1:
-1. Configure a metaflow profile. To do this, create a file named as `config_<your-metaflow-profile-name>.json` under 
+1. Configure a metaflow profile. To do this, create a file named as `config.json` under 
 `~/.metaflowconfig/`) and set the required values. Required fields to run on KFP are shared below*.
 2. Create your runs on KFP using the following command template: 
 ```
-export METAFLOW_PROFILE=<your-metaflow-profile-name> && 
-python <program-name.py> [run-on-kfp|generate-kfp-yaml] (--code-url <link-to-code>)
+python <program-name.py> kfp run
 ```
 
 ##### Option 2:
@@ -33,12 +32,7 @@ You can export the required variables* individually and run the python command:
 {
     "METAFLOW_DATASTORE_SYSROOT_S3": "s3://<path-to-s3-bucket-root>",
     "METAFLOW_DEFAULT_DATASTORE": "s3",
-
-    "METAFLOW_AWS_ARN" : "...", # required to be able to read s3 data when running within KFP
-    "METAFLOW_AWS_S3_REGION" : "...", # specify s3 region being used
-
     "KFP_RUN_URL_PREFIX" : "https://kubeflow.corp.dev.zg-aip.net" # prefix of the URL preceeding the run-id to generate correct links to the generated runs on your cluster
-
     "KFP_SDK_NAMESPACE": "example_namespace"
     "KFP_SDK_USERID": "[YOUR_NAME]@email.com"
 }
@@ -50,62 +44,36 @@ in the CLI and as environment vars/config file values. Regarding the `api_namesp
 or through the CLI, and the CLI value will take precedence. However, if you don't provide it at either place, it will be defaulted to "kubeflow".
 
 
-##### Example using a METAFLOW_PROFILE named `sree` (option 1):
+##### Example using a METAFLOW_PROFILE:
 
-Config file: `config_sree.json` (to be saved under `~/.metaflowconfig`):
+Config file: `config.json` (to be saved under `~/.metaflowconfig`):
 Contents of the config file:
 ```
 {
-    "METAFLOW_DATASTORE_SYSROOT_S3": "s3://workspace-zillow-analytics-stage/aip/metaflow",
-    "METAFLOW_DEFAULT_DATASTORE": "s3",
-
-    "METAFLOW_AWS_ARN" : "arn:aws:iam::170606514770:role/dev-zestimate-role",
-    "METAFLOW_AWS_S3_REGION" : "us-west-2",
-
-    "KFP_RUN_URL_PREFIX" : "https://kubeflow.corp.dev.zg-aip.net"
-
-    "KFP_SDK_NAMESPACE": "example_namespace"
-    "KFP_SDK_USERID": "example@email.com"
+    "KFP_RUN_URL_PREFIX": "https://kubeflow.corp.dev.zg-aip.net/",
+    "KFP_SDK_NAMESPACE": "aip-example",
+    "METAFLOW_DATASTORE_SYSROOT_S3": "s3://aip-example-dev/metaflow",
+    "METAFLOW_DEFAULT_DATASTORE": "local",
+    "METAFLOW_USER": "talebz@zillowgroup.com"
 }
 ```
 
 To `run-on-kfp` using this profile:
 ```
-export METAFLOW_PROFILE=sree && 
-python 00-helloworld/hello.py run-on-kfp 
-    --experiment-name "MF-on-KFP-P2" 
-    --run-name "hello_run" 
-    --code-url https://raw.githubusercontent.com/zillow/metaflow/mf-on-kfp-2/metaflow/tutorials/00-helloworld/hello.py
-    --namespace "example_namespace"
-    --userid "example@email.com"
+python hello.py kfp run
 ```
 
 To `generate-kfp-yaml` using this profile:
 ```
-export METAFLOW_PROFILE=sree && 
-python 00-helloworld/hello.py generate-kfp-yaml
-      --code-url https://raw.githubusercontent.com/zillow/metaflow/mf-on-kfp-2/metaflow/tutorials/00-helloworld/hello.py
-      --namespace "example_namespace" --userid "example@email.com"
+python helloworld.py kfp run --yaml-only --no-s3-code-package
 ```
 
-
-#####  Example of `run-on-kfp` without configuring a profile (option 2):
-```
-export METAFLOW_AWS_ARN="arn:aws:iam::170606514770:role/dev-zestimate-role" && 
-export METAFLOW_AWS_S3_REGION="us-west-2" && 
-export METAFLOW_DATASTORE_SYSROOT_S3="s3://workspace-zillow-analytics-stage/aip/metaflow" && 
-export KFP_RUN_URL_PREFIX="https://kubeflow.corp.dev.zg-aip.net" && 
-python 00-helloworld/hello.py run-on-kfp 
-    --code-url="https://raw.githubusercontent.com/zillow/metaflow/state-integ-s3/metaflow/tutorials/00-helloworld/hello.py"
-    --namespace "example_namespace" --userid "example@email.com"
-```
 
 #### What's happening inside the step_container_op:
 
 We execute the above local orchestration commands after performing the necessary setup. The current setup includes the following:
 
-- Download the script to be run (needed as we aren't solving code packaging yet)
-- Install the modified metaflow version (from Zillow's fork of Metaflow where we are pushing our changes)
+- Copy the code package from S3 and untar it.
 - Set a KFP user
 - Run the step command
