+++ metaflow/tutorials/00-helloworld/README.md
@@ -1,0 +1,13 @@
+# Episode 00-helloworld: Metaflow says Hi!
+
+**This flow is a simple linear workflow that verifies your installation by
+printing out 'Metaflow says: Hi!' to the terminal.**
+
+#### Showcasing:
+- Basics of Metaflow.
+- Step decorator.
+
+#### To play this episode:
+1. ```cd metaflow-tutorials```
+2. ```python 00-helloworld/helloworld.py show```
+3. ```python 00-helloworld/helloworld.py run```
